+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### 3.3.0
 - You can now define your custom action's permissions in the model config's action_permissions option
+- Returning Response or Redirect objects is now possible for admin and model configs
 
 
 ### 3.2.0